+++ Lilu/Headers/kern_cpu.hpp
@@ -92,6 +92,7 @@ namespace CPUInfo {
 		CPU_MODEL_COMETLAKE_U    =  0xA6,
 		CPU_MODEL_ROCKETLAKE_S   =  0xA7, /* desktop RocketLake */
 		CPU_MODEL_TIGERLAKE_U    =  0x8C,
+		CPU_MODEL_ALDERLAKE_S    =  0x97,
 	};
 
 	/**
@@ -124,6 +125,7 @@ namespace CPUInfo {
 		CometLake,
 		RocketLake,
 		TigerLake,
+		AlderLake,
 		MaxGeneration
 	};