+++ Changelog.md
@@ -2,6 +2,7 @@ Lilu Changelog
 ==============
 #### v1.5.8
 - Fixed kernel panic on macOS 10.15 and earlier introduced in 1.5.7
+- Added Alder Lake CPU model support
 
 #### v1.5.7
 - Added address slot support for all 64-bit macOS version